feat: implement custom logo variants via TOML configuration
